Changed code in folderIsEmpty() to be easier to read
Instead of using a directory Iterator and checking to see if next()
contains something or not, I’m using the current File object and
calling its member function to check how many files and folders exist.
It doesn’t make the code faster or any shorter, but its easier to read.